+++ model_fetch/lib/src/collection/lazy_load_bloc.dart
@@ -2,9 +2,9 @@ import 'collection_bloc.dart';
 
 abstract class LazyLoadBloc<T> extends CollectionBloc<T> {
   LazyLoadBloc({
-    required super.onError,
-    super.totalLimit,
     super.clientFilters = const [],
+    super.onError,
+    super.totalLimit,
   });
 
   Future<void> loadMoreIfCan();