+++ model_fetch_firestore/lib/src/collection/firestore_live_single_document_collection_bloc.dart
@@ -14,9 +14,9 @@ class FirestoreLiveSingleDocumentCollectionBloc<T extends WithId<String>>
   LoadStatus _loadStatus = LoadStatus.loading;
 
   FirestoreLiveSingleDocumentCollectionBloc({
-    required super.onError,
     required this.documentReference,
     required this.fromIdAndMap,
+    super.onError,
   }) {
     _snapshotSubscription = documentReference.snapshots().listen(_onSnapshot);
   }